@@ -1,47 +1,49 @@
-from typing import Tuple, Optional
+import warnings
+from enum import Enum
+from typing import Optional, Tuple
 
+import numba as nb
 import numpy as np
 from scipy.ndimage import sobel
 
-WIDTH_FIRST = 'width-first'
-HEIGHT_FIRST = 'height-first'
-VALID_ORDERS = (WIDTH_FIRST, HEIGHT_FIRST)
-
-FORWARD_ENERGY = 'forward'
-BACKWARD_ENERGY = 'backward'
-VALID_ENERGY_MODES = (FORWARD_ENERGY, BACKWARD_ENERGY)
-
 DROP_MASK_ENERGY = 1e5
 KEEP_MASK_ENERGY = 1e3
+COEFFS = np.array([0.2125, 0.7154, 0.0721], dtype=np.float32)
+
+
+class OrderMode(str, Enum):
+    WIDTH_FIRST = "width-first"
+    HEIGHT_FIRST = "height-first"
+
+
+class EnergyMode(str, Enum):
+    FORWARD = "forward"
+    BACKWARD = "backward"
+
+
+def _list_enum(enum_class) -> Tuple:
+    return tuple(x.value for x in enum_class)
 
 
 def _rgb2gray(rgb: np.ndarray) -> np.ndarray:
     """Convert an RGB image to a grayscale image"""
-    coeffs = np.array([0.2125, 0.7154, 0.0721], dtype=np.float32)
-    return (rgb @ coeffs).astype(rgb.dtype)
+    return (rgb @ COEFFS).astype(rgb.dtype)
 
 
 def _get_seam_mask(src: np.ndarray, seam: np.ndarray) -> np.ndarray:
     """Convert a list of seam column indices to a mask"""
-    return ~np.eye(src.shape[1], dtype=np.bool)[seam]
+    return np.eye(src.shape[1], dtype=bool)[seam]
 
 
 def _remove_seam_mask(src: np.ndarray, seam_mask: np.ndarray) -> np.ndarray:
     """Remove a seam from the source image according to the given seam_mask"""
     if src.ndim == 3:
         h, w, c = src.shape
-        seam_mask = np.dstack([seam_mask] * c)
-        dst = src[seam_mask].reshape((h, w - 1, c))
+        seam_mask = np.broadcast_to(seam_mask[:, :, None], src.shape)
+        dst = src[~seam_mask].reshape((h, w - 1, c))
     else:
         h, w = src.shape
-        dst = src[seam_mask].reshape((h, w - 1))
-    return dst
-
-
-def _remove_seam(src: np.ndarray, seam: np.ndarray) -> np.ndarray:
-    """Remove a seam from the source image, given a list of seam columns"""
-    seam_mask = _get_seam_mask(src, seam)
-    dst = _remove_seam_mask(src, seam_mask)
+        dst = src[~seam_mask].reshape((h, w - 1))
     return dst
 
 
@@ -56,25 +58,23 @@ def _get_energy(gray: np.ndarray) -> np.ndarray:
     return energy
 
 
+@nb.njit(nb.int32[:](nb.float32[:, :]), cache=True)
 def _get_backward_seam(energy: np.ndarray) -> np.ndarray:
     """Compute the minimum vertical seam from the backward energy map"""
-    assert energy.size > 0 and energy.ndim == 2
     h, w = energy.shape
-    cost = energy[0]
+    inf = np.array([np.inf], dtype=np.float32)
+    cost = np.concatenate((inf, energy[0], inf))
     parent = np.empty((h, w), dtype=np.int32)
     base_idx = np.arange(-1, w - 1, dtype=np.int32)
 
     for r in range(1, h):
-        left_shift = np.hstack((cost[1:], np.inf))
-        right_shift = np.hstack((np.inf, cost[:-1]))
-        min_idx = np.argmin([right_shift, cost, left_shift],
-                            axis=0) + base_idx
+        choices = np.vstack((cost[:-2], cost[1:-1], cost[2:]))
+        min_idx = np.argmin(choices, axis=0) + base_idx
         parent[r] = min_idx
-        cost = cost[min_idx] + energy[r]
+        cost[1:-1] = cost[1:-1][min_idx] + energy[r]
 
-    c = np.argmin(cost)
+    c = np.argmin(cost[1:-1])
     seam = np.empty(h, dtype=np.int32)
-
     for r in range(h - 1, -1, -1):
         seam[r] = c
         c = parent[r, c]
@@ -82,25 +82,26 @@ def _get_backward_seam(energy: np.ndarray) -> np.ndarray:
     return seam
 
 
-def _get_backward_seams(gray: np.ndarray, num_seams: int,
-                        keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _get_backward_seams(
+    gray: np.ndarray, num_seams: int, aux_energy: Optional[np.ndarray]
+) -> np.ndarray:
     """Compute the minimum N vertical seams using backward energy"""
     h, w = gray.shape
-    seams_mask = np.zeros((h, w), dtype=np.bool)
-    rows = np.arange(0, h, dtype=np.int32)
-    idx_map = np.tile(np.arange(0, w, dtype=np.int32), h).reshape((h, w))
+    seams = np.zeros((h, w), dtype=bool)
+    rows = np.arange(h, dtype=np.int32)
+    idx_map = np.broadcast_to(np.arange(w, dtype=np.int32), (h, w))
     energy = _get_energy(gray)
+    if aux_energy is not None:
+        energy += aux_energy
     for _ in range(num_seams):
-        if keep_mask is not None:
-            energy[keep_mask] += KEEP_MASK_ENERGY
         seam = _get_backward_seam(energy)
-        seams_mask[rows, idx_map[rows, seam]] = True
+        seams[rows, idx_map[rows, seam]] = True
 
         seam_mask = _get_seam_mask(gray, seam)
         gray = _remove_seam_mask(gray, seam_mask)
         idx_map = _remove_seam_mask(idx_map, seam_mask)
-        if keep_mask is not None:
-            keep_mask = _remove_seam_mask(keep_mask, seam_mask)
+        if aux_energy is not None:
+            aux_energy = _remove_seam_mask(aux_energy, seam_mask)
 
         # Only need to re-compute the energy in the bounding box of the seam
         _, cur_w = energy.shape
@@ -108,51 +109,58 @@ def _get_backward_seams(gray: np.ndarray, num_seams: int,
         hi = min(cur_w, np.max(seam) + 1)
         pad_lo = 1 if lo > 0 else 0
         pad_hi = 1 if hi < cur_w - 1 else 0
-        mid_block = gray[:, lo - pad_lo:hi + pad_hi]
+        mid_block = gray[:, lo - pad_lo : hi + pad_hi]
         _, mid_w = mid_block.shape
-        mid_energy = _get_energy(mid_block)[:, pad_lo:mid_w - pad_hi]
-        energy = np.hstack((energy[:, :lo], mid_energy, energy[:, hi + 1:]))
+        mid_energy = _get_energy(mid_block)[:, pad_lo : mid_w - pad_hi]
+        if aux_energy is not None:
+            mid_energy += aux_energy[:, lo:hi]
+        energy = np.hstack((energy[:, :lo], mid_energy, energy[:, hi + 1 :]))
 
-    return seams_mask
+    return seams
 
 
-def _get_forward_seam(gray: np.ndarray,
-                      keep_mask: Optional[np.ndarray]) -> np.ndarray:
+@nb.njit(
+    [
+        nb.int32[:](nb.float32[:, :], nb.none),
+        nb.int32[:](nb.float32[:, :], nb.float32[:, :]),
+    ],
+    cache=True,
+)
+def _get_forward_seam(gray: np.ndarray, aux_energy: Optional[np.ndarray]) -> np.ndarray:
     """Compute the minimum vertical seam using forward energy"""
-    assert gray.size > 0 and gray.ndim == 2
-    gray = gray.astype(np.float32)
     h, w = gray.shape
+    gray = np.hstack((gray[:, :1], gray, gray[:, -1:]))
+    energy = aux_energy if aux_energy is not None else np.zeros((h, w), dtype=np.float32)
 
-    top_row = gray[0]
-    top_row_lshift = np.hstack((top_row[1:], top_row[-1]))
-    top_row_rshift = np.hstack((top_row[0], top_row[:-1]))
-    dp = np.abs(top_row_lshift - top_row_rshift)
+    dp = np.empty(w + 2, dtype=np.float32)
+    dp[0] = np.inf
+    dp[1:-1] = np.abs(gray[0, 2:] - gray[0, :-2])
+    dp[-1] = np.inf
 
-    parent = np.zeros(gray.shape, dtype=np.int32)
+    parent = np.empty((h, w), dtype=np.int32)
+    choices = np.empty((3, w), dtype=np.float32)
     base_idx = np.arange(-1, w - 1, dtype=np.int32)
+    min_idx = np.empty(w, dtype=np.int32)
 
     for r in range(1, h):
-        curr_row = gray[r]
-        curr_lshift = np.hstack((curr_row[1:], curr_row[-1]))
-        curr_rshift = np.hstack((curr_row[0], curr_row[:-1]))
-        cost_top = np.abs(curr_lshift - curr_rshift)
-        if keep_mask is not None:
-            cost_top[keep_mask[r]] += KEEP_MASK_ENERGY
+        for j in range(w):
+            cmid = abs(gray[r, j + 2] - gray[r, j]) + energy[r, j]
+            choices[0, j] = cmid + abs(gray[r - 1, j + 1] - gray[r, j]) + dp[j]
+            choices[1, j] = cmid + dp[j + 1]
+            choices[2, j] = cmid + abs(gray[r - 1, j + 1] - gray[r, j + 2]) + dp[j + 2]
 
-        prev_row = gray[r - 1]
-        cost_left = cost_top + np.abs(prev_row - curr_rshift)
-        cost_right = cost_top + np.abs(prev_row - curr_lshift)
+        for j in range(w):
+            min_idx[j] = 0
+            dp[j + 1] = choices[0, j]
 
-        dp_left = np.hstack((np.inf, dp[:-1]))
-        dp_right = np.hstack((dp[1:], np.inf))
+            for i in range(1, 3):
+                if choices[i, j] < dp[j + 1]:
+                    min_idx[j] = i
+                    dp[j + 1] = choices[i, j]
 
-        choices = np.vstack([cost_left + dp_left, cost_top + dp,
-                             cost_right + dp_right])
-        dp = np.min(choices, axis=0)
-        parent[r] = np.argmin(choices, axis=0) + base_idx
+        parent[r] = min_idx + base_idx
 
-    c = np.argmin(dp)
-
+    c = np.argmin(dp[1:-1])
     seam = np.empty(h, dtype=np.int32)
     for r in range(h - 1, -1, -1):
         seam[r] = c
@@ -161,139 +169,208 @@ def _get_forward_seam(gray: np.ndarray,
     return seam
 
 
-def _get_forward_seams(gray: np.ndarray, num_seams: int,
-                       keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _get_forward_seams(
+    gray: np.ndarray, num_seams: int, aux_energy: Optional[np.ndarray]
+) -> np.ndarray:
     """Compute minimum N vertical seams using forward energy"""
     h, w = gray.shape
-    seams_mask = np.zeros((h, w), dtype=np.bool)
-    rows = np.arange(0, h, dtype=np.int32)
-    idx_map = np.tile(np.arange(0, w, dtype=np.int32), h).reshape((h, w))
+    seams = np.zeros((h, w), dtype=bool)
+    rows = np.arange(h, dtype=np.int32)
+    idx_map = np.broadcast_to(np.arange(w, dtype=np.int32), (h, w))
     for _ in range(num_seams):
-        seam = _get_forward_seam(gray, keep_mask)
-        seams_mask[rows, idx_map[rows, seam]] = True
+        seam = _get_forward_seam(gray, aux_energy)
+        seams[rows, idx_map[rows, seam]] = True
         seam_mask = _get_seam_mask(gray, seam)
         gray = _remove_seam_mask(gray, seam_mask)
         idx_map = _remove_seam_mask(idx_map, seam_mask)
-        if keep_mask is not None:
-            keep_mask = _remove_seam_mask(keep_mask, seam_mask)
+        if aux_energy is not None:
+            aux_energy = _remove_seam_mask(aux_energy, seam_mask)
 
-    return seams_mask
+    return seams
 
 
-def _get_seams(gray: np.ndarray, num_seams: int, energy_mode: str,
-               keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _get_seams(
+    gray: np.ndarray, num_seams: int, energy_mode: str, aux_energy: Optional[np.ndarray]
+) -> np.ndarray:
     """Get the minimum N seams from the grayscale image"""
-    assert energy_mode in VALID_ENERGY_MODES
-    if energy_mode == BACKWARD_ENERGY:
-        return _get_backward_seams(gray, num_seams, keep_mask)
+    gray = np.asarray(gray, dtype=np.float32)
+    if energy_mode == EnergyMode.BACKWARD:
+        return _get_backward_seams(gray, num_seams, aux_energy)
+    elif energy_mode == EnergyMode.FORWARD:
+        return _get_forward_seams(gray, num_seams, aux_energy)
     else:
-        return _get_forward_seams(gray, num_seams, keep_mask)
+        raise ValueError(
+            f"expect energy_mode to be one of {_list_enum(EnergyMode)}, got {energy_mode}"
+        )
 
 
-def _reduce_width(src: np.ndarray, delta_width: int, energy_mode: str,
-                  keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _reduce_width(
+    src: np.ndarray,
+    delta_width: int,
+    energy_mode: str,
+    aux_energy: Optional[np.ndarray],
+) -> Tuple[np.ndarray, Optional[np.ndarray]]:
     """Reduce the width of image by delta_width pixels"""
     assert src.ndim in (2, 3) and delta_width >= 0
     if src.ndim == 2:
         gray = src
         src_h, src_w = src.shape
-        dst_shape = (src_h, src_w - delta_width)
+        dst_shape: Tuple[int, ...] = (src_h, src_w - delta_width)
     else:
         gray = _rgb2gray(src)
         src_h, src_w, src_c = src.shape
         dst_shape = (src_h, src_w - delta_width, src_c)
 
-    seams_mask = _get_seams(gray, delta_width, energy_mode, keep_mask)
-    dst = src[~seams_mask].reshape(dst_shape)
-    return dst
+    to_keep = ~_get_seams(gray, delta_width, energy_mode, aux_energy)
+    dst = src[to_keep].reshape(dst_shape)
+    if aux_energy is not None:
+        aux_energy = aux_energy[to_keep].reshape(dst_shape[:2])
+    return dst, aux_energy
 
 
-def _expand_width(src: np.ndarray, delta_width: int, energy_mode: str,
-                  keep_mask: Optional[np.ndarray]) -> np.ndarray:
-    """Expand the width of image by delta_width pixels"""
-    assert src.ndim in (2, 3) and delta_width >= 0
-    if src.ndim == 2:
-        gray = src
-        src_h, src_w = src.shape
-        dst_shape = (src_h, src_w + delta_width)
-    else:
-        gray = _rgb2gray(src)
-        src_h, src_w, src_c = src.shape
-        dst_shape = (src_h, src_w + delta_width, src_c)
-
-    seams_mask = _get_seams(gray, delta_width, energy_mode, keep_mask)
-    dst = np.empty(dst_shape, dtype=np.uint8)
-
+@nb.njit(
+    nb.float32[:, :, :](nb.float32[:, :, :], nb.boolean[:, :], nb.int32), cache=True
+)
+def _insert_seams_kernel(
+    src: np.ndarray, seams: np.ndarray, delta_width: int
+) -> np.ndarray:
+    """The numba kernel for inserting seams"""
+    src_h, src_w, src_c = src.shape
+    dst = np.empty((src_h, src_w + delta_width, src_c), dtype=src.dtype)
     for row in range(src_h):
         dst_col = 0
         for src_col in range(src_w):
-            if seams_mask[row, src_col]:
-                lo = max(0, src_col - 1)
-                hi = src_col + 1
-                dst[row, dst_col] = src[row, lo:hi].mean(axis=0)
+            if seams[row, src_col]:
+                left = src[row, max(src_col - 1, 0)]
+                right = src[row, src_col]
+                dst[row, dst_col] = (left + right) / 2
                 dst_col += 1
             dst[row, dst_col] = src[row, src_col]
             dst_col += 1
-        assert dst_col == src_w + delta_width
-
     return dst
 
 
-def _resize_width(src: np.ndarray, width: int, energy_mode: str,
-                  keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _insert_seams(src: np.ndarray, seams: np.ndarray, delta_width: int) -> np.ndarray:
+    """Insert multiple seams into the source image"""
+    dst = src.astype(np.float32)
+    if dst.ndim == 2:
+        dst = dst[:, :, None]
+    dst = _insert_seams_kernel(dst, seams, delta_width).astype(src.dtype)
+    if src.ndim == 2:
+        dst = dst.squeeze(-1)
+    return dst
+
+
+def _expand_width(
+    src: np.ndarray,
+    delta_width: int,
+    energy_mode: str,
+    aux_energy: Optional[np.ndarray],
+    step_ratio: float,
+) -> Tuple[np.ndarray, Optional[np.ndarray]]:
+    """Expand the width of image by delta_width pixels"""
+    assert src.ndim in (2, 3) and delta_width >= 0
+    if not 0 < step_ratio <= 1:
+        raise ValueError(f"expect `step_ratio` to be between (0,1], got {step_ratio}")
+
+    dst = src
+    while delta_width > 0:
+        max_step_size = max(1, round(step_ratio * dst.shape[1]))
+        step_size = min(max_step_size, delta_width)
+        gray = dst if dst.ndim == 2 else _rgb2gray(dst)
+        seams = _get_seams(gray, step_size, energy_mode, aux_energy)
+        dst = _insert_seams(dst, seams, step_size)
+        if aux_energy is not None:
+            aux_energy = _insert_seams(aux_energy, seams, step_size)
+        delta_width -= step_size
+
+    return dst, aux_energy
+
+
+def _resize_width(
+    src: np.ndarray,
+    width: int,
+    energy_mode: str,
+    aux_energy: Optional[np.ndarray],
+    step_ratio: float,
+) -> Tuple[np.ndarray, Optional[np.ndarray]]:
     """Resize the width of image by removing vertical seams"""
     assert src.size > 0 and src.ndim in (2, 3)
     assert width > 0
-    assert energy_mode in VALID_ENERGY_MODES
 
     src_w = src.shape[1]
     if src_w < width:
-        dst = _expand_width(src, width - src_w, energy_mode, keep_mask)
+        dst, aux_energy = _expand_width(
+            src, width - src_w, energy_mode, aux_energy, step_ratio
+        )
     else:
-        dst = _reduce_width(src, src_w - width, energy_mode, keep_mask)
+        dst, aux_energy = _reduce_width(src, src_w - width, energy_mode, aux_energy)
+    return dst, aux_energy
+
+
+def _transpose_image(src: np.ndarray) -> np.ndarray:
+    """Transpose a source image in rgb or grayscale format"""
+    if src.ndim == 3:
+        dst = src.transpose((1, 0, 2))
+    else:
+        dst = src.T
     return dst
 
 
-def _resize_height(src: np.ndarray, height: int, energy_mode: str,
-                   keep_mask: Optional[np.ndarray]) -> np.ndarray:
+def _resize_height(
+    src: np.ndarray,
+    height: int,
+    energy_mode: str,
+    aux_energy: Optional[np.ndarray],
+    step_ratio: float,
+) -> Tuple[np.ndarray, Optional[np.ndarray]]:
     """Resize the height of image by removing horizontal seams"""
     assert src.ndim in (2, 3) and height > 0
-    if src.ndim == 3:
-        src = _resize_width(src.transpose((1, 0, 2)), height, energy_mode,
-                            keep_mask).transpose((1, 0, 2))
-    else:
-        src = _resize_width(src.T, height, energy_mode, keep_mask).T
-    return src
+    if aux_energy is not None:
+        aux_energy = aux_energy.T
+    src = _transpose_image(src)
+    src, aux_energy = _resize_width(src, height, energy_mode, aux_energy, step_ratio)
+    src = _transpose_image(src)
+    if aux_energy is not None:
+        aux_energy = aux_energy.T
+    return src, aux_energy
 
 
-def _check_mask(mask: np.ndarray, shape: Tuple[int, int]) -> np.ndarray:
+def _check_mask(mask: np.ndarray, shape: Tuple[int, ...]) -> np.ndarray:
     """Ensure the mask to be a 2D grayscale map of specific shape"""
-    mask = np.asarray(mask, dtype=np.bool)
+    mask = np.asarray(mask, dtype=bool)
     if mask.ndim != 2:
-        raise ValueError('Invalid mask of shape {}: expected to be a 2D '
-                         'binary map'.format(mask.shape))
+        raise ValueError(f"expect mask to be a 2d binary map, got shape {mask.shape}")
     if mask.shape != shape:
-        raise ValueError('The shape of mask must match the image: expected {}, '
-                         'got {}'.format(shape, mask.shape))
+        raise ValueError(
+            f"expect the shape of mask to match the image, got {mask.shape} vs {shape}"
+        )
     return mask
 
 
 def _check_src(src: np.ndarray) -> np.ndarray:
     """Ensure the source to be RGB or grayscale"""
-    src = np.asarray(src, dtype=np.uint8)
+    src = np.asarray(src)
     if src.size == 0 or src.ndim not in (2, 3):
-        raise ValueError('Invalid src of shape {}: expected an 3D RGB image or '
-                         'a 2D grayscale image'.format(src.shape))
+        raise ValueError(
+            f"expect a 3d rgb image or a 2d grayscale image, got image in shape {src.shape}"
+        )
     return src
 
 
-def resize(src: np.ndarray, size: Tuple[int, int],
-           energy_mode: str = 'backward', order: str = 'width-first',
-           keep_mask: Optional[np.ndarray] = None) -> np.ndarray:
+def resize(
+    src: np.ndarray,
+    size: Optional[Tuple[int, int]] = None,
+    energy_mode: str = "backward",
+    order: str = "width-first",
+    keep_mask: Optional[np.ndarray] = None,
+    drop_mask: Optional[np.ndarray] = None,
+    step_ratio: float = 0.5,
+) -> np.ndarray:
     """Resize the image using the content-aware seam-carving algorithm.
 
     :param src: A source image in RGB or grayscale format.
-    :param size: The target size in pixels, as a 2-tuple (width, height)
+    :param size: The target size in pixels, as a 2-tuple (width, height).
     :param energy_mode: Policy to compute energy for the source image. Could be
         one of ``backward`` or ``forward``. If ``backward``, compute the energy
         as the gradient at each pixel. If ``forward``, compute the energy as the
@@ -303,47 +380,78 @@ def resize(src: np.ndarray, size: Tuple[int, int],
         remove or insert all vertical seams first, then the horizontal ones,
         while ``height-first`` is the opposite.
     :param keep_mask: An optional mask where the foreground is protected from
-        seam removal. If not specify, no area will be protected.
+        seam removal. If not specified, no area will be protected.
+    :param drop_mask: An optional binary object mask to remove. If given, the
+        object will be removed before resizing the image to the target size.
+    :param step_ratio: The maximum size expansion ratio in one seam carving step.
+        The image will be expanded in multiple steps if target size is too large.
     :return: A resized copy of the source image.
     """
     src = _check_src(src)
-    src_h, src_w = src.shape[:2]
 
-    width, height = size
-    width = int(round(width))
-    height = int(round(height))
-    if width <= 0 or height <= 0:
-        raise ValueError('Invalid size {}: expected > 0'.format(size))
-    if width >= 2 * src_w:
-        raise ValueError('Invalid target width {}: expected less than twice '
-                         'the source width (< {})'.format(width, 2 * src_w))
-    if height >= 2 * src_h:
-        raise ValueError('Invalid target height {}: expected less than twice '
-                         'the source height (< {})'.format(height, 2 * src_h))
+    if order not in _list_enum(OrderMode):
+        raise ValueError(
+            f"expect order to be one of {_list_enum(OrderMode)}, got {order}"
+        )
 
-    if order not in VALID_ORDERS:
-        raise ValueError('Invalid order {}: expected {}'.format(
-            order, VALID_ORDERS))
-
-    if energy_mode not in VALID_ENERGY_MODES:
-        raise ValueError('Invalid energy mode {}: expected {}'.format(
-            energy_mode, VALID_ENERGY_MODES))
+    aux_energy = None
 
     if keep_mask is not None:
-        keep_mask = _check_mask(keep_mask, (src_h, src_w))
+        keep_mask = _check_mask(keep_mask, src.shape[:2])
 
-    if order == WIDTH_FIRST:
-        src = _resize_width(src, width, energy_mode, keep_mask)
-        src = _resize_height(src, height, energy_mode, keep_mask)
-    else:
-        src = _resize_height(src, height, energy_mode, keep_mask)
-        src = _resize_width(src, width, energy_mode, keep_mask)
+        aux_energy = np.zeros(src.shape[:2], dtype=np.float32)
+        aux_energy[keep_mask] += KEEP_MASK_ENERGY
+
+    # remove object if `drop_mask` is given
+    if drop_mask is not None:
+        drop_mask = _check_mask(drop_mask, src.shape[:2])
+
+        if aux_energy is None:
+            aux_energy = np.zeros(src.shape[:2], dtype=np.float32)
+        aux_energy[drop_mask] -= DROP_MASK_ENERGY
+
+        if order == OrderMode.HEIGHT_FIRST:
+            src = _transpose_image(src)
+            aux_energy = aux_energy.T
+
+        num_seams = (aux_energy < 0).sum(1).max()
+        while num_seams > 0:
+            src, aux_energy = _reduce_width(src, num_seams, energy_mode, aux_energy)
+            num_seams = (aux_energy < 0).sum(1).max()
+
+        if order == OrderMode.HEIGHT_FIRST:
+            src = _transpose_image(src)
+            aux_energy = aux_energy.T
+
+    # resize image if `size` is given
+    if size is not None:
+        width, height = size
+        width = round(width)
+        height = round(height)
+        if width <= 0 or height <= 0:
+            raise ValueError(f"expect target size to be positive, got {size}")
+
+        if order == OrderMode.WIDTH_FIRST:
+            src, aux_energy = _resize_width(
+                src, width, energy_mode, aux_energy, step_ratio
+            )
+            src, aux_energy = _resize_height(
+                src, height, energy_mode, aux_energy, step_ratio
+            )
+        else:
+            src, aux_energy = _resize_height(
+                src, height, energy_mode, aux_energy, step_ratio
+            )
+            src, aux_energy = _resize_width(
+                src, width, energy_mode, aux_energy, step_ratio
+            )
 
     return src
 
 
-def remove_object(src: np.ndarray, drop_mask: np.ndarray,
-                  keep_mask: Optional[np.ndarray] = None) -> np.ndarray:
+def remove_object(
+    src: np.ndarray, drop_mask: np.ndarray, keep_mask: Optional[np.ndarray] = None
+) -> np.ndarray:
     """Remove an object on the source image.
 
     :param src: A source image in RGB or grayscale format.
@@ -352,6 +460,12 @@ def remove_object(src: np.ndarray, drop_mask: np.ndarray,
         removal. If not specified, no area is protected.
     :return: A copy of the source image where the drop_mask is removed.
     """
+    warnings.warn(
+        "`remove_object` is deprecated in favor of `resize(src, drop_mask=mask)`, and will be removed in the next version of seam-carving",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+
     src = _check_src(src)
 
     drop_mask = _check_mask(drop_mask, src.shape[:2])
