@@ -1,7 +1,4 @@
-from seam_carving.carve import resize, remove_object
+from seam_carving.carve import remove_object, resize
 
-__version__ = '1.0.0'
-__all__ = [
-    'resize',
-    'remove_object'
-]
+__version__ = "1.1.0"
+__all__ = ["resize", "remove_object"]
